Extract decidim_controllers entrypoint
