+++ decidim-assemblies/app/packs/src/decidim/assemblies/admin/assemblies.js
@@ -1,4 +1,5 @@
 import { definitionsFromContext } from "src/decidim/refactor/support/stimulus"
-
-const context = require.context("../controllers", true, /controller\.js$/)
-window.Stimulus.load(definitionsFromContext(context))
+document.addEventListener("DOMContentLoaded", () => {
+  const context = require.context("../controllers", true, /controller\.js$/)
+  window.Stimulus.load(definitionsFromContext(context))
+});
