+++ decidim-core/app/packs/src/decidim/index.js
@@ -260,14 +260,3 @@ document.addEventListener("comments:loaded", (event) => {
   }
 });
 
-import { Application } from "@hotwired/stimulus"
-import { definitionsFromContext } from "src/decidim/refactor/support/stimulus"
-
-const application = Application.start()
-application.debug = true
-
-const context = require.context("./controllers", true, /controller\.js$/)
-application.load(definitionsFromContext(context))
-
-window.definitionsFromContext = definitionsFromContext
-window.Stimulus = application
