+++ decidim-core/app/packs/src/decidim/refactor/support/stimulus.js
@@ -5,8 +5,12 @@
   In order to have our own structure, we need to replace the Regex from identifierForContextKey method
  */
 
-const identifierForContextKey = function(key) {
+const identifierForContextKey = function(key, excludes = []) {
   const logicalName = (key.match(/^(?:\.\/)?(.+)(?:\/controller\..+?)$/) || [])[1];
+
+  if (logicalName && excludes.includes(logicalName)) {
+    return null;
+  }
 
   if (logicalName) {
     return logicalName.replace(/_/g, "-").replace(/\//g, "--");
@@ -22,17 +26,18 @@ const definitionForModuleAndIdentifier = function(module, identifier) {
   return null;
 }
 
-const definitionForModuleWithContextAndKey = function(context, key) {
-  const identifier = identifierForContextKey(key);
+const definitionForModuleWithContextAndKey = function(context, key, excludes = []) {
+  const identifier = identifierForContextKey(key, excludes);
+
   if (identifier) {
     return definitionForModuleAndIdentifier(context(key), identifier);
   }
   return null;
 }
 
-const definitionsFromContext = function(context) {
+const definitionsFromContext = function(context, excludes = []) {
   return context.keys().
-    map((key) => definitionForModuleWithContextAndKey(context, key)).
+    map((key) => definitionForModuleWithContextAndKey(context, key, excludes)).
     filter((value) => value);
 }
