+++ decidim-core/app/views/layouts/decidim/_decidim_javascript.html.erb
@@ -1,9 +1,9 @@
 <%# NOTE: for performance reasons defer this javascript call would be good, although there are references to jquery that need to be reviewed %>
 
 <% if Decidim.service_worker_enabled %>
-  <%= javascript_pack_tag "decidim_core","decidim_sw", defer: false %>
+  <%= javascript_pack_tag "decidim_core", "decidim_controllers", "decidim_sw", defer: false %>
 <% else %>
-  <%= javascript_pack_tag "decidim_core", defer: false %>
+  <%= javascript_pack_tag "decidim_core", "decidim_controllers", defer: false %>
 <% end %>
 
 <%= render partial: "layouts/decidim/js_configuration" %>
