+++ decidim-core/config/assets.rb
@@ -4,6 +4,7 @@ base_path = File.expand_path("..", __dir__)
 
 Decidim::Shakapacker.register_path("#{base_path}/app/packs")
 Decidim::Shakapacker.register_entrypoints(
+  decidim_controllers: "#{base_path}/app/packs/entrypoints/decidim_controllers.js",
   decidim_overrides: "#{base_path}/app/packs/entrypoints/decidim_overrides.scss",
   decidim_core: "#{base_path}/app/packs/entrypoints/decidim_core.js",
   decidim_sw: "#{base_path}/app/packs/entrypoints/decidim_sw.js",
