+++ decidim-core/spec/lib/shakapacker/runner_spec.rb
@@ -32,6 +32,7 @@ module Shakapacker
         expect(runtime_config["default"]["additional_paths"]).to include("app/packs")
         expect(runtime_config["default"]["additional_paths"]).to include("#{core_path}/app/packs")
         expect(runtime_config["default"]["entrypoints"]).to include(
+          "decidim_controllers" => "#{core_path}/app/packs/entrypoints/decidim_controllers.js",
           "decidim_core" => "#{core_path}/app/packs/entrypoints/decidim_core.js",
           "decidim_sw" => "#{core_path}/app/packs/entrypoints/decidim_sw.js",
           "decidim_conference_diploma" => "#{core_path}/app/packs/entrypoints/decidim_conference_diploma.js",
