+++ decidim-core/spec/system/data_consent/scripts_spec.rb
@@ -28,7 +28,7 @@ describe "Data consent scripts" do
           <title>Accessibility Test</title>
           #{stylesheet_pack_tag "decidim_core"}
           #{stylesheet_pack_tag "decidim_dev"}
-          #{javascript_pack_tag "decidim_core", "decidim_dev", defer: false}
+          #{javascript_pack_tag "decidim_core", "decidim_controllers", "decidim_dev", defer: false}
         </head>
         <!-- Add some line breaks so that the "WAI WCAG" notification does not block screenshots -->
         <br><br><br><br><br>
