+++ decidim-design/app/views/layouts/decidim/design/_head.html.erb
@@ -1,5 +1,5 @@
 <%= tag :meta, name: "viewport", content: "width=device-width, initial-scale=1.0" %>
-<%= javascript_pack_tag "decidim_core", "decidim_design", defer: false %>
+<%= javascript_pack_tag "decidim_core", "decidim_controllers", "decidim_design", defer: false %>
 
 <%= append_stylesheet_pack_tag "decidim_overrides", media: "all" %>
 <%= append_stylesheet_pack_tag "decidim_design", media: "all" %>
