+++ decidim-meetings/app/packs/src/decidim/meetings/meetings.js
@@ -2,5 +2,7 @@ import "src/decidim/meetings/meetings_form"
 import "src/decidim/meetings/meetings_polls"
 import { definitionsFromContext } from "src/decidim/refactor/support/stimulus"
 
-const context = require.context("./controllers", true, /controller\.js$/)
-window.Stimulus.load(definitionsFromContext(context))
+document.addEventListener("DOMContentLoaded", () => {
+  const context = require.context("./controllers", true, /controller\.js$/)
+  window.Stimulus.load(definitionsFromContext(context))
+});
