+++ decidim-system/app/views/layouts/decidim/system/_header.html.erb
@@ -2,4 +2,4 @@
 <%= csrf_meta_tags %>
 <%= append_stylesheet_pack_tag "decidim_system_overrides", media: "all" %>
 <%= stylesheet_pack_tag "decidim_core", "decidim_system", media: "all" %>
-<%= javascript_pack_tag "decidim_core", "decidim_admin", "decidim_system", defer: false %>
+<%= javascript_pack_tag "decidim_core", "decidim_controllers", "decidim_admin", "decidim_system", defer: false %>
